ai_tribal_marketplace/backend/requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.29.0

sqlalchemy==2.0.29

pillow==10.3.0

deep-translator==1.11.4

torch>=2.2.0
transformers>=4.40.0

python-multipart==0.0.9
